refactor(types): Dependency* DTO 下沉到 src/types/ 解开 types → eval-core 反向依赖
src/types/doctor.ts:2 之前 type-only import `DependencyRequirements` 来自
eval-core/dependency-checker,而 eval-core/dependency-checker.ts:12 又正向
import types/index —— types 作为底层契约层不应反向看实现模块。

将 DependencyRequirements / DependencyReasonCode / DependencyIssue /
DependencyCheckResult 4 个纯 DTO 上移到 src/types/dependencies.ts(types/
barrel 同步 export);eval-core/dependency-checker.ts 改为从 types/ import
并 re-export(对外签名不变,所有现有 consumer 继续 work)。

types/ 现在严格单向被消费,无 schema 改动。
