+++ src/eval-core/dependency-checker.ts
@@ -9,41 +9,20 @@
 import { existsSync } from 'node:fs';
 import { delimiter, dirname, join, resolve } from 'node:path';
 import { execFileSync, execSync } from 'node:child_process';
-import type { Artifact, Sample } from '../types/index.js';
+import type {
+  Artifact,
+  DependencyCheckResult,
+  DependencyIssue,
+  DependencyRequirements,
+  Sample,
+} from '../types/index.js';
 
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-export interface DependencyRequirements {
-  tools?: string[];
-  files?: string[];
-  env?: string[];
-  preflight?: string[];
-}
-
-export type DependencyReasonCode =
-  | 'tool_not_found'
-  | 'file_not_found'
-  | 'env_not_set'
-  | 'preflight_failed';
-
-export interface DependencyIssue {
-  category: 'tool' | 'file' | 'env' | 'preflight';
-  name: string;
-  /** Stable enum so consumers (doctor, eval-pipeline) translate per their own lang.
-   *  The dep-checker itself never produces user-facing strings — it only carries facts. */
-  reasonCode: DependencyReasonCode;
-  /** Optional untranslated raw context: stderr line, cwd, command — surfaces at the
-   *  consumer's UI layer so the actual failure is visible without round-tripping the
-   *  process. Never localized; pass through verbatim. */
-  reasonDetail?: string;
-}
-
-export interface DependencyCheckResult {
-  ok: boolean;
-  missing: DependencyIssue[];
-}
+export type {
+  DependencyCheckResult,
+  DependencyIssue,
+  DependencyReasonCode,
+  DependencyRequirements,
+} from '../types/index.js';
 
 // ---------------------------------------------------------------------------
 // Extraction — scan text for dependencies
